refactor: Place type 분리
