+++ src/components/domain/CourseSlider/index.tsx
@@ -4,10 +4,10 @@ import 'slick-carousel/slick/slick-theme.css';
 import styled from '@emotion/styled';
 import theme from '~/styles/theme';
 import CourseSliderItem from './CourseSliderItem';
-import { IPlace } from '~/pages/course/[id]';
+import { PlaceData } from '~/types/place';
 
 interface CourseSliderProps {
-  places?: IPlace[];
+  places?: PlaceData[];
 }
 
 const CourseSlider = ({ places }: CourseSliderProps) => {
