+++ src/components/domain/Map/CourseMap/index.tsx
@@ -3,19 +3,8 @@ import { Map, MapMarker, CustomOverlayMap, Polyline } from 'react-kakao-maps-sdk
 import markerIcon from 'public/assets/place/course.png';
 import Script from 'next/script';
 import styled from '@emotion/styled';
+import { IPlace } from '~/types/place';
 
-export interface IPlace {
-  id: number;
-  name: string;
-  description: string;
-  address: string;
-  latitude: string;
-  longitude: string;
-  phoneNumber: string;
-  imageUrl: string;
-  isRecommended: boolean;
-  isThumbnail: boolean;
-}
 interface CourseType {
   id: number;
   latitude: string;
