+++ src/types/course.ts
@@ -1,3 +1,5 @@
+import { IPlace } from './place';
+
 export interface ICourseItem {
   id: number;
   title: string;
@@ -10,6 +12,25 @@ export interface ICourseItem {
   isBookmarked: boolean;
   nickname: string;
   profileUrl: string;
+}
+
+export interface ICourseDetail {
+  id: number;
+  title: string;
+  thumbnail: string;
+  region: string;
+  period: string;
+  themes: string[];
+  spots: string[];
+  places: IPlace[];
+  likes: number;
+  isLiked: boolean;
+  isBookmarked: boolean;
+  nickname: string;
+  userId: number;
+  profileImage: string;
+  createdAt: string;
+  updatedAt: string;
 }
 
 export interface CourseFilter {
