+++ src/types/place.ts
@@ -7,3 +7,16 @@ export interface IPlaceItem {
   thumbnail: string;
   bookmarked: boolean;
 }
+
+export interface IPlace {
+  id: number;
+  name: string;
+  description: string;
+  address: string;
+  latitude: string;
+  longitude: string;
+  phoneNumber: string;
+  imageUrl: string;
+  isRecommended: boolean;
+  isThumbnail: boolean;
+}
